+++ src/cpp/runtime/esp_idf/device_config_store.h
@@ -14,7 +14,7 @@
 
 #include "esp_err.h"
 #include "runtime/espectre_protocol.h"
-#include "runtime/runtime_interface.h"
+#include "runtime/runtime_config.h"
 
 namespace espectre {