+++ src/cpp/runtime/runtime_config_utils.h
@@ -9,7 +9,8 @@
  */
 #pragma once
 
-#include "runtime_interface.h"
+#include "runtime_config.h"
+#include "runtime_snapshot.h"
 
 namespace espectre {